third_party/WebKit/LayoutTests/http/tests/inspector/bindings/bindings-multiple-frames.html

<head>
<script src="../inspector-test.js"></script>
<script src="./bindings-test.js"></script>

<script>

async function test()
{
    InspectorTest.markStep('initialWorkspace');
    var snapshot = InspectorTest.dumpWorkspace();

    InspectorTest.markStep('createIframes');
    await Promise.all([
        InspectorTest.attachFrame('frame1', './resources/magic-frame.html', '_test_create-iframe1.js'),
        InspectorTest.attachFrame('frame2', './resources/magic-frame.html', '_test_create-iframe2.js'),
    ]);
    snapshot = InspectorTest.dumpWorkspace(snapshot);

    InspectorTest.markStep('detachFrame1');
    await InspectorTest.detachFrame('frame1', '_test_detachFrame1.js');
    snapshot = InspectorTest.dumpWorkspace(snapshot);

    InspectorTest.markStep('detachFrame2');
    await InspectorTest.detachFrame('frame2', '_test_detachFrame2.js');
    snapshot = InspectorTest.dumpWorkspace(snapshot);

    InspectorTest.completeTest();
}

</script>
</head>
<body onload="runTest()">
<p>
Verify that UISourceCodes are removed as iframes are getting detached.
</p>
</body>
