third_party/WebKit/LayoutTests/http/tests/inspector/bindings/bindings-multiple-frames-expected.txt

Verify that UISourceCodes are removed as iframes are getting detached.


Running: initialWorkspace
Workspace: 7 uiSourceCodes.
    debugger:///VM[XXX] bindings-multiple-frames.html
    debugger:///VM[XXX] bindings-multiple-frames.html
    debugger:///VM[XXX] bindings-test.js
    debugger:///VM[XXX] inspector-test.js
    http://127.0.0.1:8000/inspector/bindings/bindings-multiple-frames.html
    http://127.0.0.1:8000/inspector/bindings/bindings-test.js
    http://127.0.0.1:8000/inspector/inspector-test.js

Running: createIframes
Workspace: 19 uiSourceCodes.
    debugger:///VM[XXX] bindings-multiple-frames.html
    debugger:///VM[XXX] bindings-multiple-frames.html
    debugger:///VM[XXX] bindings-test.js
    debugger:///VM[XXX] inspector-test.js
    debugger:///VM[XXX] magic-script.js
    debugger:///VM[XXX] magic-script.js
    debugger:///VM[XXX] _test_create-iframe1.js
    debugger:///VM[XXX] _test_create-iframe2.js
    http://127.0.0.1:8000/inspector/bindings/bindings-multiple-frames.html
    http://127.0.0.1:8000/inspector/bindings/bindings-test.js
    http://127.0.0.1:8000/inspector/bindings/resources/magic-frame.html
    http://127.0.0.1:8000/inspector/bindings/resources/magic-frame.html
    http://127.0.0.1:8000/inspector/bindings/resources/magic-script.js
    http://127.0.0.1:8000/inspector/bindings/resources/magic-script.js
    http://127.0.0.1:8000/inspector/bindings/resources/magic-style.css
    http://127.0.0.1:8000/inspector/bindings/resources/magic-style.css
    http://127.0.0.1:8000/inspector/inspector-test.js
    _test_create-iframe1.js
    _test_create-iframe2.js

Running: detachFrame1
Workspace: 18 uiSourceCodes.
    debugger:///VM[XXX] bindings-multiple-frames.html
    debugger:///VM[XXX] bindings-multiple-frames.html
    debugger:///VM[XXX] bindings-test.js
    debugger:///VM[XXX] inspector-test.js
    debugger:///VM[XXX] magic-script.js
    debugger:///VM[XXX] magic-script.js
    debugger:///VM[XXX] _test_create-iframe1.js
    debugger:///VM[XXX] _test_create-iframe2.js
    debugger:///VM[XXX] _test_detachFrame1.js
    http://127.0.0.1:8000/inspector/bindings/bindings-multiple-frames.html
    http://127.0.0.1:8000/inspector/bindings/bindings-test.js
    http://127.0.0.1:8000/inspector/bindings/resources/magic-frame.html
    http://127.0.0.1:8000/inspector/bindings/resources/magic-script.js
    http://127.0.0.1:8000/inspector/bindings/resources/magic-style.css
    http://127.0.0.1:8000/inspector/inspector-test.js
    _test_create-iframe1.js
    _test_create-iframe2.js
    _test_detachFrame1.js

Running: detachFrame2
Workspace: 17 uiSourceCodes.
    debugger:///VM[XXX] bindings-multiple-frames.html
    debugger:///VM[XXX] bindings-multiple-frames.html
    debugger:///VM[XXX] bindings-test.js
    debugger:///VM[XXX] inspector-test.js
    debugger:///VM[XXX] magic-script.js
    debugger:///VM[XXX] magic-script.js
    debugger:///VM[XXX] _test_create-iframe1.js
    debugger:///VM[XXX] _test_create-iframe2.js
    debugger:///VM[XXX] _test_detachFrame1.js
    debugger:///VM[XXX] _test_detachFrame2.js
    http://127.0.0.1:8000/inspector/bindings/bindings-multiple-frames.html
    http://127.0.0.1:8000/inspector/bindings/bindings-test.js
    http://127.0.0.1:8000/inspector/inspector-test.js
    _test_create-iframe1.js
    _test_create-iframe2.js
    _test_detachFrame1.js
    _test_detachFrame2.js
